Actually navigate between Login and MainTabs
