+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
+    "@react-navigation/native-stack": "^7.2.0",
     "@reduxjs/toolkit": "^2.5.1",
     "react": "18.3.1",
     "react-native": "0.77.0",
