+++ src/config/routes.ts
@@ -1,0 +1,1 @@
+
