+++ src/navigators/RootNavigator.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import {TabNavigator} from './TabNavigator';
+import {LoginScreen} from '../screens/LoginScreen';
+
+export type RootStackParamList = {
+  Login: undefined;
+  MainTabs: undefined;
+};
+
+const Stack = createNativeStackNavigator<RootStackParamList>();
+
+export const RootNavigator = () => {
+  return (
+    <Stack.Navigator initialRouteName="Login" screenOptions={{headerShown: false}}>
+      <Stack.Screen name="Login" component={LoginScreen} />
+      <Stack.Screen name="MainTabs" component={TabNavigator} />
+    </Stack.Navigator>
+  );
+};
